pyproject.toml
[project]
name = "image-caption-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "clearml>=2.1.2",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "nltk>=3.9.2",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
]
